+++ CHANGELOG.md
@@ -2,7 +2,17 @@
 
 All notable changes to the HumanKind Funeral Notices plugin will be documented in this file.
 
-## [2.2.17] - 2025-10-01
+## [2.3.0] - 2025-10-02
+
+### 📊 **Anonymous Analytics System**
+
+#### **Added**
+- ✨ **Anonymous Usage Analytics** - Plugin now collects anonymous usage statistics to improve features and provide industry insights
+- 📍 **Default Venue/Location** - Sites can now set a default venue that auto-populates for new funeral notices
+
+---
+
+## [2.2.16] - 2025-10-01
 
 ### 🔄 **Update System Improvements**
 
@@ -16,7 +26,7 @@ All notable changes to the HumanKind Funeral Notices plugin will be documented i
 
 ---
 
-## [2.2.16] - 2025-10-01
+## [2.2.15] - 2025-10-01
 
 ### 🔧 **SEO & Configuration Fixes**